current tools (with Bugfix)
ptybuffer and socklinger had a possible buffer overrun

However this was on the socket name used, not the peer name.
Hence unlikely to be exploited.

I use these myself a lot, and no application used some
exploitable variant, because usually you want your sockets
to be predefined.  Especially in root context.